Update DataScienceManifesto.md
This suggestion is about rephrasing "The Data scientist is honest" section. I think "obvious" is a word that we should use rarely in a manifesto, as it highly depends on the reader's experiences. Then I tried to clarify on what problems other data scientists face getting a model from another data scientist.